refactor: Group entities in less .ts files
Don't a file for each entity to reduce boilerplate
