+++ src/components/EditLaboratory/LaboratoryDetails.tsx
@@ -2,8 +2,8 @@ import { EditLaboratoryContext } from "@/context/laboratories/EditLaboratoryCont
 import { EditLaboratoryActionType } from "@/hooks/laboratories/editLaboratoryTypes";
 import { updateLaboratoryDetailsService } from "@/services/laboratories/update-laboratory-details.service";
 import { getTeacherRubricsService } from "@/services/rubrics/get-teacher-rubrics.service";
-import { LaboratoryBaseInfo } from "@/types/entities/laboratory";
-import { CreatedRubric } from "@/types/entities/rubric";
+import { LaboratoryBaseInfo } from "@/types/entities/laboratory-entities";
+import { CreatedRubric } from "@/types/entities/rubric-entities";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { Save } from "lucide-react";
 import { useContext, useEffect, useState } from "react";
