+++ src/services/laboratories/get-course-laboratories.service.ts
@@ -1,4 +1,4 @@
-import { LaboratoryBaseInfo } from "@/types/entities/laboratory";
+import { LaboratoryBaseInfo } from "@/types/entities/laboratory-entities";
 import { AxiosError } from "axios";
 
 import { GenericResponse, HttpRequester } from "../axios";
