+++ src/services/laboratories/get-laboratory-by-uuid.service.ts
@@ -1,4 +1,8 @@
-import { Laboratory, MarkdownBlock } from "@/types/entities/laboratory";
+import {
+  Laboratory,
+  MarkdownBlock,
+  TestBlock
+} from "@/types/entities/laboratory-entities";
 import { AxiosError } from "axios";
 
 import { GenericResponse, HttpRequester } from "../axios";
@@ -38,15 +42,17 @@ export const getLaboratoryByUUIDService = async (
     laboratory.blocks.push(...markdownBlocks);
 
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const testBlocks: any[] = data.test_blocks.map((block: any) => ({
+    const testBlocks: TestBlock[] = data.test_blocks.map((block: any) => ({
       uuid: block.uuid,
       languageUUID: block.language_uuid,
       testArchiveUUID: block.test_archive_uuid,
+      submissionUUID: block.submission_uuid,
       name: block.name,
-      index: block.index,
-      blockType: "test"
+      index: block.index
     }));
     laboratory.blocks.push(...testBlocks);
+
+    console.table(testBlocks);
 
     // Sort blocks by index
     laboratory.blocks.sort((a, b) => a.index - b.index);
