--- src/types/entities/course.ts
@@ -1,5 +1,0 @@
-export type Course = {
-  uuid: string;
-  name: string;
-  color: string;
-};
