--- src/types/entities/enrolled-student.ts
@@ -1,6 +1,0 @@
-export type EnrolledStudent = {
-  uuid: string;
-  institutional_id: string;
-  full_name: string;
-  is_active: boolean;
-};
