--- src/types/entities/student.ts
@@ -1,5 +1,0 @@
-export type Student = {
-  uuid: string;
-  institutional_id: string;
-  full_name: string;
-};
